revert strict param changes and add folder
